+++ src/Detach/Utils/VectorUtils.cs
@@ -4,14 +4,25 @@ namespace Detach.Utils;
 
 public static class VectorUtils
 {
-	public static float GetAngle(Vector2 a, Vector2 b)
+	public static float GetAngleFrom(Vector2 vector)
+	{
+		return MathF.Atan2(vector.Y, vector.X);
+	}
+
+	/// <summary>
+	/// Returns the angle in radians between two directional vectors.
+	/// </summary>
+	public static float GetAngleBetween(Vector2 a, Vector2 b)
 	{
 		float dotProduct = a.X * b.X + a.Y * b.Y;
 		float determinant = a.X * b.Y - a.Y * b.X;
 		return MathF.Atan2(determinant, dotProduct);
 	}
 
-	public static float GetAngle(Vector3 a, Vector3 b)
+	/// <summary>
+	/// Returns the angle in radians between two directional vectors.
+	/// </summary>
+	public static float GetAngleBetween(Vector3 a, Vector3 b)
 	{
 		float dotProduct = Vector3.Dot(a, b);
 		float determinant = Vector3.Cross(a, b).Length();